Bug 945222 - Add test harness for running W3C web-platform-tests.
This adds a new test harness, and an associated
mach command |mach web-platform-tests|.

At present only desktop firefox is supported. Tests
are run in parallel using multiple firefox instances.

It also adds a copy of the web platform tests and a
(incomplete) script for updating them.
* * *
Bug 945222 - Update wptrunner to use proxy autoconfig and localhost in a way that might work on infrastructure.
* * *
Change test runner manager to be a bit more robust?